Refactor to app router and fix build
