+++ app/inventory/page.tsx
