+++ app/manufacturing/page.tsx
