+++ middleware.ts
@@ -1,4 +1,4 @@
-export { auth as middleware } from "next-auth/middleware"
+export { auth as middleware } from '@/auth'
 
 export const config = {
   matcher: [
