+++ package.json
@@ -7,7 +7,8 @@
     "build": "prisma generate && next build",
     "start": "next start -p 3000",
     "postinstall": "prisma generate",
-    "seed": "tsx prisma/seed.ts"
+    "seed": "tsx prisma/seed.ts",
+    "test": "echo \"No tests specified\" && exit 0"
   },
   "dependencies": {
     "@prisma/client": "^5.16.1",
