+++ seed.ts
@@ -1,4 +1,4 @@
-import { prisma } from '../lib/prisma'
+import { prisma } from './lib/prisma'
 import { hash } from 'bcryptjs'
 
 async function main() {
